attack norm done
