+++ include/cub3d_bonus.h
@@ -132,6 +132,8 @@ typedef struct s_render
 {
 	double	transform_x;
 	double	transform_y;
+	double	sprite_x;
+	double	sprite_y;
 	int		tex_x;
 	int		draw_end_y;
 	int		less_height;
@@ -395,7 +397,8 @@ double	get_delta_time(void);
 void	create_fireball(t_game *game);
 void	update_fireballs(t_game *game);
 void	render_fireballs(t_game *game, t_attack *fireball);
-void	calculate_fireball_position(t_game *game, t_attack *fireball);
+void	calculate_fireball_position(t_game *game, t_attack *fireball,
+		mlx_texture_t *texture);
 
 // update_game
 void	update_game_state(t_game *game);
@@ -414,7 +417,8 @@ void	fire_projectile(t_game *game);
 void	update_bullets(t_game *game);
 void	boss_move(t_game *game, double dx, double dy);
 void	calculate_boss_position(t_game *game, double pos_x, double pos_y,
-			mlx_texture_t *texture, int less_height);
-void	calculate_bullet_position(t_game *game, t_attack *bullet);
+		mlx_texture_t *texture, int less_height);
+void	calculate_bullet_position(t_game *game, t_attack *bullet,
+		mlx_texture_t *texture);
 
 #endif
