+++ src/bonus/game_bonus/handle_game_bonus.c
@@ -16,20 +16,18 @@ static void	calculate_potion_position(t_game *game, double pos_x, double pos_y,
 	mlx_texture_t *texture)
 {
 	t_render	*rend;
-	double		sprite_x;
-	double		sprite_y;
 	double		inv_det;
 
 	rend = ft_calloc(1, sizeof(t_render));
 	rend->less_height = 12000;
-	sprite_x = pos_x - game->player->pos_x;
-	sprite_y = pos_y - game->player->pos_y;
+	rend->sprite_x = pos_x - game->player->pos_x;
+	rend->sprite_y = pos_y - game->player->pos_y;
 	inv_det = 1.0 / (game->player->camera_dir_x * game->player->player_dir_y
 			- game->player->player_dir_x * game->player->camera_dir_y);
-	rend->transform_x = inv_det * (game->player->player_dir_y * sprite_x
-			- game->player->player_dir_x * sprite_y);
-	rend->transform_y = inv_det * (-game->player->camera_dir_y * sprite_x
-			+ game->player->camera_dir_x * sprite_y);
+	rend->transform_x = inv_det * (game->player->player_dir_y * rend->sprite_x
+			- game->player->player_dir_x * rend->sprite_y);
+	rend->transform_y = inv_det * (-game->player->camera_dir_y * rend->sprite_x
+			+ game->player->camera_dir_x * rend->sprite_y);
 	if (rend->transform_y <= 0)
 		return ;
 	rend->sprite_height = (int)fabs(HEIGHT / rend->transform_y);
@@ -44,20 +42,18 @@ static void	calculate_key_position(t_game *game, double pos_x, double pos_y,
 	mlx_texture_t *texture)
 {
 	t_render	*rend;
-	double		sprite_x;
-	double		sprite_y;
 	double		inv_det;
 
 	rend = ft_calloc(1, sizeof(t_render));
 	rend->less_height = 10000;
-	sprite_x = pos_x - game->player->pos_x;
-	sprite_y = pos_y - game->player->pos_y;
+	rend->sprite_x = pos_x - game->player->pos_x;
+	rend->sprite_y = pos_y - game->player->pos_y;
 	inv_det = 1.0 / (game->player->camera_dir_x * game->player->player_dir_y
 			- game->player->player_dir_x * game->player->camera_dir_y);
-	rend->transform_x = inv_det * (game->player->player_dir_y * sprite_x
-			- game->player->player_dir_x * sprite_y);
-	rend->transform_y = inv_det * (-game->player->camera_dir_y * sprite_x
-			+ game->player->camera_dir_x * sprite_y);
+	rend->transform_x = inv_det * (game->player->player_dir_y * rend->sprite_x
+			- game->player->player_dir_x * rend->sprite_y);
+	rend->transform_y = inv_det * (-game->player->camera_dir_y * rend->sprite_x
+			+ game->player->camera_dir_x * rend->sprite_y);
 	if (rend->transform_y <= 0)
 		return ;
 	rend->sprite_height = (int)fabs(HEIGHT / rend->transform_y);
